+++ bindings/c/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Added
+
+- `keep_at_rules` option [#265](https://github.com/Stranger6667/css-inline/issues/265)
+
 ## [0.16.0] - 2025-07-16
 
 ### Changed